+++ c/libmcpb/include/mcpb/mcpb.h
@@ -43,9 +43,9 @@ typedef enum
 const char *mcpb_strerror(int err);
 
 #define MCPB_VERSION_MAJOR 0
-#define MCPB_VERSION_MINOR 2
-#define MCPB_VERSION_PATCH 1
-#define MCPB_VERSION_STRING "0.2.1"
+#define MCPB_VERSION_MINOR 3
+#define MCPB_VERSION_PATCH 0
+#define MCPB_VERSION_STRING "0.3.0"
 
 #ifdef __cplusplus
 } /* extern "C" */